Un-hardcode message ID in test backend
`EmailMultiAlternatives.message()` serializes an email message for
sending, and among other things adds a unique "Message-ID" header if one
isn't already present.  Use this in place of a hard-coded message ID in
the test backend.

For tests that inspect the message ID in the `anymail_status`, provide
an explicit "Message-ID" header rather than using the ephemeral
generated one.